+++ packages/bloc/lib/bloc.dart
@@ -1,3 +1,4 @@
 library bloc;
 
 export './src/bloc.dart';
+export './src/transition.dart';